Ajout d'un dernier paramètre de configuration du plugin:
le répertoire local du développeur, qui sera utilisé par le plugin
maven, comme repo git local pour exécuter le git push du war à déployer.
Ce répertoire doit être non-utilisé par l'IDE (Eclipse IntelliJ etc...),
afin que celui-ci ne soit pas pertubé par l'action du plugin maven.

Il peut être configuré dans le pom.xml de l"appli web utilisatrice:
<repertoire-repo-git-local>cheminDuRepertoire</repertoire-repo-git-local>
Vous y siasirez le chemin du répertoire, comme vous le saisissez
nativemetn pour l'OS de votre machine de dev

+++ deployeur/src/main/java/lasselle/deployeur/JiblMojo1.java
@@ -123,7 +123,9 @@ public class JiblMojo1 extends AbstractMojo {
 		
 		
 	}
-	
+//	@Parameter(alias = "repertoire-repo-git-local", property = "repertoire-repo-git-local", defaultValue = "${project.build.directory}/tempmvnautobuild")
+	@Parameter(alias = "repertoire-repo-git-local", property = "repertoire-repo-git-local", required=true)
+	String cheminRepoGitLocalDeTravail = null;
 	private void copierFihierWarVersCible() {
 //		JiblScpTo.faisCopie(this.cheminProjet + "/" + this.nomFichierWAR, this.nomFichierWAR, adresseIPcibleDeploiement, SSHusername, SSHuserpwd);
 //		JiblSftp.faisCopie(this.cheminProjet + "/" + this.nomFichierWAR, this.nomFichierWAR, adresseIPcibleDeploiement, SSHusername, SSHuserpwd);
@@ -136,11 +138,13 @@ public class JiblMojo1 extends AbstractMojo {
 		// 
 		// => Je créée un nouveau repo git, séparé, et dans C:\moi\mes_repos_git
 		File reposDIR = new File("C:\\moi\\mes_repos_git");
-		String cheminRepo = "C:\\moi\\mes_repos_git\\tempmvnautobuild";
+		String cheminRepo = this.cheminRepoGitLocalDeTravail;
 		File repoDIR = new File(cheminRepo);
 		// je le détruis, et le re-créée
 		try {
-			FileUtils.forceDelete(repoDIR);
+			if (repoDIR.exists()) {
+				FileUtils.forceDelete(repoDIR);
+			}
 		} catch (IOException e2) {
 			System.out.println(" JIBL + pb au delete initla du répertoire du repo [" + cheminRepo + "]");
 			e2.printStackTrace();
